requirements.txt
pykrx==1.0.45
pandas>=2.0.0
python-telegram-bot==21.3
python-dotenv==1.0.1
schedule==1.2.2
